+++ src/components/IndividualCard.tsx
@@ -1,6 +1,7 @@
 import styled from '@emotion/native';
 import React, {useRef} from 'react';
 import {Image} from 'react-native';
+import {ActionSheetRef} from 'react-native-actions-sheet';
 
 import MoreIcon from '../assets/icons/more-icon.svg';
 import MaleIcon from '../assets/icons/male.svg';
@@ -10,8 +11,7 @@ import {formatKorean} from '../utils/format-date';
 import {useNav} from '../hooks/useNav';
 import useIndividualIdStore from '../stores/useIndividualIdStore';
 import Button from './common/Button';
-import ActionSheet, {ActionSheetRef} from 'react-native-actions-sheet';
-import {Headline6} from './common/TextGroup';
+import IndividualMoreSeeActionSheet from './IndividualMoreSeeActionSheet';
 
 export interface IndividualCardProps {
   individual: Individual;
@@ -34,18 +34,6 @@ const Card = styled.TouchableOpacity`
   margin: 8px;
 `;
 
-const EditButtonBlock = styled.View`
-  padding-bottom: 100px;
-`;
-
-const EditButtonOuter = styled.View`
-  padding: 25px 0;
-  border-bottom-width: 1px;
-  border-bottom-color: ${props => props.theme.color.lightGray};
-`;
-
-const EditButton = styled(Button)``;
-
 const MoreButton = styled(Button)`
   position: absolute;
   top: 0px;
@@ -56,24 +44,29 @@ const IndividualBox = styled.View`
   background-color: white;
   width: 100%;
 `;
+
 const ProfileBlock = styled.View`
   display: flex;
   flex-direction: row;
   justify-content: center;
   padding: 4px;
 `;
+
 const ProfileImage = styled(Image)`
   width: 100px;
   height: 100px;
   border-radius: 10px;
 `;
+
 const InfoBlock = styled.View`
   display: flex;
   gap: 5px;
 `;
+
 const HatchDate = styled.Text`
   text-align: center;
 `;
+
 const NameAndGenderBlock = styled.View`
   display: flex;
   flex-direction: row;
@@ -82,6 +75,7 @@ const NameAndGenderBlock = styled.View`
   gap: 3px;
   text-align: center;
 `;
+
 const Name = styled.Text``;
 const Gender = styled.Text``;
 
@@ -90,6 +84,7 @@ const MemoBlock = styled.View`
   border-radius: 10px;
   background-color: #ffdfbe;
 `;
+
 const Memo = styled.Text`
   border-radius: 5px;
   text-align: center;
@@ -101,8 +96,13 @@ const IndividualCardComponent = ({individual}: IndividualCardProps) => {
   const navigation = useNav<'IndividualTopTab'>();
   const updateId = useIndividualIdStore(state => state.updateId);
   const actionSheetRef = useRef<ActionSheetRef>(null);
+
   const openActionSheet = () => actionSheetRef.current?.show();
   const closeActionSheet = () => actionSheetRef.current?.hide();
+  const moveToEdit = () => {
+    closeActionSheet();
+    navigation.push('IndividualInfoEditScreen');
+  };
 
   return (
     <Card
@@ -110,25 +110,11 @@ const IndividualCardComponent = ({individual}: IndividualCardProps) => {
         updateId(individual.id);
         navigation.push('IndividualTopTab');
       }}>
-      <ActionSheet ref={actionSheetRef}>
-        <EditButtonBlock>
-          <EditButtonOuter>
-            <EditButton
-              varient={'text'}
-              onPress={() => {
-                closeActionSheet();
-                navigation.push('IndividualInfoEditScreen');
-              }}>
-              <Headline6>수정</Headline6>
-            </EditButton>
-          </EditButtonOuter>
-          <EditButtonOuter>
-            <EditButton varient={'text'}>
-              <Headline6 color={theme.color.red}>삭제</Headline6>
-            </EditButton>
-          </EditButtonOuter>
-        </EditButtonBlock>
-      </ActionSheet>
+      <IndividualMoreSeeActionSheet
+        actionSheetRef={actionSheetRef}
+        onPressEdit={moveToEdit}
+        onPressDelete={() => {}}
+      />
       <MoreButton varient={'text'} onPress={openActionSheet}>
         <MoreIcon width={24} height={24} fill={theme.color.gray} />
       </MoreButton>
